Add the capability to convert NSTimestamp into epoch time and vice versa
This is the first batch of modification which handles NSTimestamp only.
If you set the property er.rest.timestampFormat=epoch, ERRest will
convert a NSTimestamp value into a long value (number of milliseconds
since 01/01/70).
The main difference is we have to rename
ERXRestUtils.coerceValueToString() into ERXRestUtils.convertValue() as
this method doesn’t always return a String.